+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 This Changelog only applies to notable changes of `helios.js` on the main branch.
 
+# 2022/10/25
+* Better error message when catching `RuntimeError` inside `evalParam`
+
 # 2022/10/23
 * HashedDatum and InlineDatum no longer exported
 * TxId and DatumHash exported, `Hash` no longer exported